+++ chess/eval.hpp
@@ -3,8 +3,8 @@
 
 #include "chess.hpp"
 
-int evaluate( position const& p, color::type c );
+short evaluate( position const& p, color::type c );
 
-int evaluate_move( position const& p, color::type c, int current_evaluation, move const& m );
+short evaluate_move( position const& p, color::type c, short current_evaluation, move const& m );
 
 #endif
